+++ packages/component/src/hooks/useAnimatingToEnd.js
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+
+import StateContext from '../ScrollToBottom/StateContext';
+
+export default function useAnimatingToEnd() {
+  const context = useContext(StateContext);
+
+  return [context.animatingToEnd];
+}