+++ packages/component/src/index.js
@@ -8,11 +8,13 @@ import Panel from './ScrollToBottom/Panel';
 import StateContext from './ScrollToBottom/StateContext';
 
 import useAnimating from './hooks/useAnimating';
+import useAnimatingToEnd from './hooks/useAnimatingToEnd';
 import useAtBottom from './hooks/useAtBottom';
 import useAtEnd from './hooks/useAtEnd';
 import useAtStart from './hooks/useAtStart';
 import useAtTop from './hooks/useAtTop';
 import useMode from './hooks/useMode';
+import useObserveScrollPosition from './hooks/useObserveScrollPosition';
 import useScrollTo from './hooks/useScrollTo';
 import useScrollToBottom from './hooks/useScrollToBottom';
 import useScrollToEnd from './hooks/useScrollToEnd';
@@ -29,11 +31,13 @@ export {
   Panel,
   StateContext,
   useAnimating,
+  useAnimatingToEnd,
   useAtBottom,
   useAtEnd,
   useAtStart,
   useAtTop,
   useMode,
+  useObserveScrollPosition,
   useScrollTo,
   useScrollToBottom,
   useScrollToEnd,